Refactor: Move platform-specific classes to separate files
This commit refactors the codebase by moving platform-specific classes (Android and iOS) into their respective dedicated files. This improves code organization and maintainability.

Additionally, the module version number has been incremented to 0.1.2.
