+++ gradle.properties
@@ -19,7 +19,7 @@ gpr.user=yet300
 gpr.key=todo
 
 MODULE_PACKAGE_NAME=dev.gitlive
-MODULE_VERSION_NUMBER=0.1.1
+MODULE_VERSION_NUMBER=0.1.2
 MODULE_NAME=appauth-kotlin
 
 OPEN_SOURCE_REPO=https://oss.sonatype.org/service/local/staging/deploy/maven2/
